app/retrieval/engine.py
from azure.search.documents import SearchClient
from azure.core.credentials import AzureKeyCredential

class RetrievalEngine:
    def __init__(self, endpoint: str, key: str, index_name: str):
        self.client = SearchClient(endpoint, AzureKeyCredential(key), index_name)

    def query_index(self, query: str, vector_query: Optional[list] = None):
        """Query Azure AI Search for relevant compliance documents."""
        # TODO: Implement vector and hybrid search
        pass

    def index_document(self, document: dict):
        """Upload content to Azure AI Search."""
        # TODO: Implement indexing logic
        pass
