requirements.txt
fastapi
uvicorn
langgraph
langchain
langchain-openai
azure-ai-search-documents
azure-storage-blob
azure-identity
yt-dlp
python-dotenv
pydantic
pydantic-settings
appinsights
langsmith
